refactor: imported and separated the Ctype Types for use in the demo cli
