+++ src/index.ts
@@ -12,6 +12,7 @@ import * as Balance from './balance/Balance.chain'
 import Identity from './identity/Identity'
 import PublicIdentity, { IURLResolver } from './identity/PublicIdentity'
 import CType from './ctype/CType'
+import CTypeMetadata from './ctype/CTypeMetadata'
 import * as CTypeUtils from './ctype/CTypeUtils'
 import Claim from './claim/Claim'
 import RequestForAttestation from './requestforattestation/RequestForAttestation'
@@ -32,6 +33,7 @@ export * from './errorhandling/ExtrinsicError'
 // ---- Types, which define the most basic KILT objects ----
 export { default as IPublicIdentity } from './types/PublicIdentity'
 export { default as ICType } from './types/CType'
+export { default as ICtypeMetadata } from './types/CTypeMetedata'
 export { default as IClaim } from './types/Claim'
 export { default as IAttestedClaim } from './types/AttestedClaim'
 export { default as IAttestation } from './types/Attestation'
@@ -52,6 +54,7 @@ export {
   PublicIdentity,
   IURLResolver,
   CType,
+  CTypeMetadata,
   CTypeUtils,
   Claim,
   RequestForAttestation,
@@ -69,6 +72,7 @@ export {
   CTypeModel,
   CTypeInputModel,
   CTypeWrapperModel,
+  CTypeWrapperMetadata,
 } from './ctype/CTypeSchema'
 export * from './messaging/Message'
 
@@ -79,6 +83,7 @@ export default {
   Identity,
   PublicIdentity,
   CType,
+  CTypeMetadata,
   Claim,
   RequestForAttestation,
   Attestation,
