+++ src/types/CType.ts
@@ -18,19 +18,3 @@ export default interface ICType {
   owner: IPublicIdentity['address'] | null
   schema: ICTypeSchema
 }
-
-export interface ICtypeMetadata {
-  metadata: {
-    $id: string
-    $schema: string
-    title: {
-      type: string
-    }
-    description: {
-      type: string
-    }
-    properties: object
-    type: 'object'
-  }
-  ctypeHash: string
-}
